+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 * Add `ovlpr replace`
 * Add `ovlpr restrict`
 
+### Changed
+
+* Github Actions publish.yml
+
 ## [0.4.1] - 2019-09-10
 
 ### Added